Merge pull request #2775 from RosettaCommons/vmullig/indexselector_noerror
Add an option to the ResidueIndexSelector to allow it to ignore residues that aren't in the pose

Currently, the `ResidueIndexSelector` throws an error if a residue index specified does not exist in the pose (_e.g._ residue 56 of a 55-residue pose).  While this is a good default behaviour, users have encountered situations in which this is not a desired behaviour.  This pull request retains the current behaviour as default behaviour, but adds an option to disable this.

Tasks:
- [x] Add option to ResidueIndexSelector.
- [x] Setters & getters.
- [x] Update `parse_my_tag()`.
- [x] Update xsd functions.
- [x] Update unit test.
- [x] Update serialization functions.
- [x] Beauty.
- [x] Documentation.

@srgerb

Old repository SHA1: 48774481a3fe653f2096390263417bcff339c76a