Merge pull request #2824 from RosettaCommons/jadolfbr/glycosylation_pdbinfo
Keep input PDB numbering post-glycosylation

This PR finally fixes pose glycosylation to keep the input PDBInfo in tact and create sensible PDBInfo records for the new glycan. 

The new glycan is now part of the Protein chain it was added to.  The resulting Pose has the same numbering as before.  This will streamline the use of the `SimpleGlycosylateMover` and glycosylation code for structure solutions, especially for EM density. 

This is on by default, but can be turned off.  All integration tests that glycosylate and output PDBs will change. 

Old repository SHA1: 5385b4d95ff17bd55e06dc3610abf82cf327b27f