Merge pull request #2887 from RosettaCommons/revert-2886-revert-2883-vmullig/rotamer_read_refactor
Re-merge refactor of the rotamer library reader.

This was originally merged as pull request #2883.  When I tested that PR, the Intel icc build was broken; the fix to that unfortunately broke the gcc build, and I didn't catch that before merging.  I reverted that merge, and will fix the issue here and re-merge.

Reverts RosettaCommons/main#2886 (i.e. this is a reversion of the reversion of the initial merge, reinstating the initial merge).  This should be merged before merging #2855.

Old repository SHA1: 90e7110210c63e393b0d98a2bbbd17f6487a51ee