Merge pull request #2904 from RosettaCommons/JackMaguire/MCHBNetCorrection
MC HBNet Correction. There was a small bug in the MC HBNet code that did not impact the legitimacy of results but strongly hindered the sampling ability of the MC HBNet protocol (credit goes to Scott for discovering there was a bug). That bug is now corrected and I took this opportunity to clean up the code in the class surrounding the bug.

Old repository SHA1: d1c71d6b842cbf58a55f60e02b7f55bd5a3f143e