Merge pull request #2935 from RosettaCommons/vmullig/disable_rotamer_sorting
Adding a means of skipping the rotamer well order correction.

So it turns out that, now that I've implemented the Voronoi-based rotamer well interpolation for noncanonicals, it probably doesn't actually matter what order rotamer wells are in. The sorting step, which takes a while for large rotamer libraries, can therefore be skipped. I'm keeping the option to turn it back on in case there are unanticipated cases in which rotamer well order matters, but setting the default behaviour to be no sorting.

Old repository SHA1: be9fbaab46287e27ac087b388dae4bc441cf2d47