Merge pull request #2972 from RosettaCommons/JackMaguire/MultistageRosettaScripts
There was a bug in an integration test I added to master yesterday that causes it to always fail. I did not notice it because the testing server did not have a reference to "diff" against, but now that other people are pushing to master I can see that all the diffs are backtraces. My apologies to everyone who had to deal with these useless test failures!

Old repository SHA1: 29d49b005cfec62b16526fc7a454f10051fc7f05