Merge pull request #2966 from RosettaCommons/JackMaguire/MCHBNetCorrection
This PR introduces an option to HBNet that allows HBNet's internal scores to be placed in the output poses (turned off by default). This would allow the user to access the scores using the CalculatorFilter or ReadPoseExtraScoreFilter and the scores would be included in the score.sc file.

Old repository SHA1: fc64acf459d38b80f23e187a17c9d44d4060ee21