Merge pull request #2982 from RosettaCommons/jadolfbr/config
Support for common Flag Config files

This PR adds the ability for rosetta to find and read common flag config (fconfig) files.  This allows us to always run rosetta with a set of custom common options.  

If a file named `common` is found in `$HOME/.rosetta/flags` or if it/they are in the current working directory, we use that instead. You can set any number of flag configurations with the `-fconfig` option By default, we have:
```
   -fconfig common
```

 This is useful if you have a set of flags for different purposes - like design, glycans, and antibody, so you could do something like:

```
  -fconfig common antibody 
```

 That would load both the common and antibody configurations. 


 If you have a common flag file (which is loaded by default if present), you skip loading through an option
```
  -no_config  
```

* Note, `-no_fconfig` is used for both unit and integration tests.  

Old repository SHA1: 529d335a5417819986236595a131a4166c5c2149