Merge pull request #3057 from RosettaCommons/vmullig/fix_integration_test_using_score12
Switching rotamer_recovery integration test back to score12.

The rotamer_recovery integration test used to use the -correct flag. I took this out when I added an error message / failure for using -correct with talaris or later, since the test was not set to use a pre-talaris scorefunction (i.e. it was using ref2015 with -correct). However, this has revealed aberrant behaviour that was being masked by the spurious use of the -correct flag. Since the test is clearly testing something that's not supported post-score12, and since I don't know the original intent of the test, I'm switching it back to use score12 for now.

@everyday847 @aleaverfay

Old repository SHA1: 7ae1b4433846007b5bdaba3bf29460cd359c6481