Merge pull request #3023 from RosettaCommons/vmullig/add_metal_csts
Adding `metalbinding_constraint` to weights files in database that have constraints turned on

The -auto_setup_metals flag turns on the metalbinding_constraint score term in the default scorefunction, but if a user specifies a custom scorefunction (or uses Sharon's very useful SetupMetalsMover in RosettaScripts with a script-defined scorefunction), this is not on by default. To make it a bit easier for users, I think it makes sense to turn on the metalbinding_constraint term in all of the scorefunctions that are defined with *_cst.wts files in the database. (All other constraint types are on in these scorefunctions.)

Integration test changes are due to extra column in score line in tests that use the *_cst.wts files.  This is exactly the expected behaviour.  Merging.

Old repository SHA1: 023bf2be7f93cff335d82769912a4cd4cb91af3f