Merge pull request #3098 from RosettaCommons/vmullig/fix_nterm_proline_error
Fix bug in simple_cycpep_predict with N-terminal prolines and isopeptide bonds.

The issue was that the app is trying to sample the cis conformation of the peptide bond preceding the proline, which it can't do in the isopeptide case with proline at the N-terminus.  Fixed.

Old repository SHA1: bfdb5e490e47e1efd380dc51f8950f94082df1f7