Merge pull request #3135 from RosettaCommons/vmullig/buried_unsat_score_incremental_merge_1
Incremental merge of buried unsatisfied penalty scoreterm.  (This currently lacks symmetry support.)

This is an incremental merge of pull request #2597.  This pull request adds basic support for asymmetric poses, but no support yet for symmetric poses.  This scoreterm guides the packer to solutions with no buried unsatisfied hydrogen bond donors or acceptors.  It is intended for use with the rest of the scorefunction, and possibly for use with the `voids_penalty`, `aa_composition`, and `hbnet` scoreterms.

Tasks:
- [x] Basic support for asymmetric poses.
- [x] (Temporary) error if used on a symmetric pose.
- [x] Unit tests.
- [x] Integration tests.
- [x] Documentation.
- [x] Beauty.

Old repository SHA1: 83cbd28d003c13365cdf0e6d48421adf0203f744