Merge pull request #3180 from RosettaCommons/roccomoretti/fix_pymol_cif
(Partially) Fix from-Pymol CIF reading

Pymol's mmCIF output is apparently less extensive than from-RCSB ones, and this is causing the Rosetta CIF reader to choke. 

Note that Pymol CIF input still has an issue - apparently CIFPARSE-OBJ (the external library from the RCSB which we use to parse CIF files) completely ignores the last table in the CIF file. For from-Pymol output this is the atom_site table, which means that you lose all the coordinate data, resulting in an empty pose. The way to fix it is for you to add a dummy line to the end of the CIF file. As I'm not keen on debugging the external library, I've added a warning message to highlight how to get around the issue.

Old repository SHA1: 25a6ce563ed91d7d3f9122eb57b986f4bb7cbed2

Co-authored-by: Andy Watkins <andy.watkins2@gmail.com>