Merge pull request #3231 from CyrusBiotechnology/231-forcommons-options-order-test
Unit test to ensure that the order in which options are interpreted it as-intended and remains stable. This checks a vector option to make sure it appends and a single-valued option to make sure the "last" one always dominates.

Old repository SHA1: fe7d6edbed85cfbf21488775ed8a164ed3d2f565