Merge pull request #3233 from RosettaCommons/jadolfbr/cart_glycan_min
Add per-atom Cartesian Minimization Control

This PR adds the ability to control specific atoms in cartesian minimization through a new `set_atom( AtomID, bool`) function and associated map in the `MoveMap`

Additionally, a boolean option is added to the `MoveMapFactory`, `cartesian`, in order to allow us to set specific behavior for `cartesian` vs `dihedral` movemaps.  This option is available in code or through RosettaScripts, and currently is only in use for glycans. It could be expanded in the future for lipids or other polymeric systems. 

This PR addresses issue #3181, allowing cartesian minimization of glycans to match dihedral minimization and standard IUPAC nomenclature of residues and torsions.  

This PR will also allow refinement of specific glycan residues in cartesian space, or per-atom refinement for ligands. However, since we do not yet have an `AtomSelector`, this functionality is only currently accessible via code. 



Unit tests for the `MoveMapFactory` and glycan-specific movemap creation have been added. An integration test for the full functionality has also been added.   

Finally, this fixes GlycanRelax and GlycanTreeRelax to use the new settings for cartesian.  

Old repository SHA1: d21730674410d87996e8a77de75365626ca9ce8f