Merge pull request #3239 from RosettaCommons/BYachnin/enzdes_preverse_header_errorhandling
The enzdes constraint machinery relies on a set of REMARK headers in the
pdb file. If the -run:preserve header option is NOT set, these remarks
get discarded when Rosetta reads in the PDB. If the user tries to apply
enzdes constraints, the error message that is produced indicates that
there are no REMARKs in the PDB file.

I have added a quick check to see if -run:preserve_header is set, and if
not, add additional tracer output to inform the user that they probably
need to turn it on.

Old repository SHA1: 3b2286a7b511fd5d81d43e80afbeb0889f606983